+++ lib/mix/tasks/scrypath.failed.ex
@@ -16,7 +16,10 @@ defmodule Mix.Tasks.Scrypath.Failed do
     {opts, argv} = OperatorTask.parse!(args)
     schema = OperatorTask.schema_from_argv!(argv)
 
-    case Scrypath.failed_sync_work(schema, OperatorTask.runtime_opts(opts) ++ OperatorTask.test_operator_opts()) do
+    case Scrypath.failed_sync_work(
+           schema,
+           OperatorTask.runtime_opts(opts) ++ OperatorTask.test_operator_opts()
+         ) do
       {:ok, failed_work} -> Mix.shell().info(OperatorTask.render_failed_work(schema, failed_work))
       {:error, reason} -> OperatorTask.error!("scrypath.failed", reason)
     end