+++ lib/mix/tasks/scrypath.status.ex
@@ -16,7 +16,10 @@ defmodule Mix.Tasks.Scrypath.Status do
     {opts, argv} = OperatorTask.parse!(args)
     schema = OperatorTask.schema_from_argv!(argv)
 
-    case Scrypath.sync_status(schema, OperatorTask.runtime_opts(opts) ++ OperatorTask.test_operator_opts()) do
+    case Scrypath.sync_status(
+           schema,
+           OperatorTask.runtime_opts(opts) ++ OperatorTask.test_operator_opts()
+         ) do
       {:ok, status} -> Mix.shell().info(OperatorTask.render_status(status))
       {:error, reason} -> OperatorTask.error!("scrypath.status", reason)
     end